requirement.txt
PyJWT==1.5.0
django-braces==1.11.0
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.6
oauthlib==2.0.1
python-openid==2.2.5
requests-oauthlib==0.8.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
